training/requirements.txt
edgeimpulse
tensorflow-rocm
numpy<2
pandas
matplotlib
scikit-learn
